+++ src/service/SubtasksHandler.java
@@ -1,37 +1,30 @@
 package service;
 
 import com.sun.net.httpserver.HttpExchange;
-import com.sun.net.httpserver.HttpHandler;
 import objects.Subtask;
 
 import java.io.IOException;
 import java.util.Optional;
 
-public class SubtasksHandler extends BaseHttpHandler implements HttpHandler {
-    private final TaskManager taskManager;
+public class SubtasksHandler extends BaseHttpHandler {
 
     public SubtasksHandler(TaskManager taskManager) {
-        this.taskManager = taskManager;
+        super(taskManager); // Явный вызов конструктора родителя
     }
 
     @Override
     public void handle(HttpExchange exchange) throws IOException {
         try {
             String method = exchange.getRequestMethod();
-            String path = exchange.getRequestURI().getPath();
 
-            switch (method) {
-                case "GET":
-                    handleGet(exchange);
-                    break;
-                case "POST":
-                    handlePost(exchange);
-                    break;
-                case "DELETE":
-                    handleDelete(exchange);
-                    break;
-                default:
-                    sendText(exchange, "Метод не поддерживается", 405);
+            if ("GET".equals(method)) {
+                handleGet(exchange);
+            } else if ("POST".equals(method)) {
+                handlePost(exchange);
+            } else if ("DELETE".equals(method)) {
+                handleDelete(exchange);
+            } else {
+                sendText(exchange, "Метод не поддерживается", 405);
             }
         } catch (Exception e) {
             sendInternalError(exchange);
@@ -76,13 +69,8 @@ public class SubtasksHandler extends BaseHttpHandler implements HttpHandler {
     private void handleDelete(HttpExchange exchange) throws IOException {
         Optional<Integer> idOpt = getPathId(exchange);
         if (idOpt.isPresent()) {
-            Subtask subtask = taskManager.getSubtask(idOpt.get());
-            if (subtask != null) {
-                taskManager.deleteSubtask(idOpt.get());
-                sendText(exchange, "Подзадача удалена");
-            } else {
-                sendNotFound(exchange);
-            }
+            taskManager.deleteSubtask(idOpt.get());
+            sendText(exchange, "Подзадача удалена");
         } else {
             taskManager.deleteAllSubtasks();
             sendText(exchange, "Все подзадачи удалены");
